feat: move components into a single file
