--- libs/shared/common/src/lib/components/datepicker/calendar/calendar-year-view/calendar-year-view.component.html
@@ -1,48 +1,0 @@
-<div class="grid gap-4">
-  <div class="flex items-center justify-between">
-    <div class="flex items-center justify-center">
-      <span class="font-bold cursor-default">{{ view() | date: 'YYYY' }}</span>
-    </div>
-    <div class="flex items-center gap-1">
-      <div class="flex items-center">
-        <button class="p-2 et-button" (click)="changeYear(-1)" [expensesTrackerTooltip]="previousYearTooltip">
-          <span class="material-icons">chevron_left</span>
-        </button>
-        <expenses-tracker-tooltip #previousYearTooltip>
-          <span>Go to the previous month</span>
-        </expenses-tracker-tooltip>
-      </div>
-      <div class="flex items-center">
-        <button class="p-2 et-button" (click)="resetDate()" [expensesTrackerTooltip]="resetTooltip">
-          <span class="material-icons">radio_button_checked</span>
-        </button>
-        <expenses-tracker-tooltip #resetTooltip>
-          <span>Go to today</span>
-        </expenses-tracker-tooltip>
-      </div>
-      <div class="flex items-center">
-        <button class="p-2 et-button" (click)="changeYear(1)" [expensesTrackerTooltip]="nextYearTooltip">
-          <span class="material-icons">chevron_right</span>
-        </button>
-        <expenses-tracker-tooltip #nextYearTooltip>
-          <span>Go to the next month</span>
-        </expenses-tracker-tooltip>
-      </div>
-    </div>
-  </div>
-  <div class="grid gap-1 grid-cols-4 grid-rows-3">
-    <ng-container *ngFor="let month of monthsOfYear">
-      <button
-        class="h-full p-2 et-button"
-        [ngClass]="{
-          'bg-color-accent-fg text-color-fg-on-emphasis': isSelectedMonth(month),
-          'bg-color-transparent': !isSelectedMonth(month)
-        }"
-        (click)="selectMonth(month)">
-        <span class="font-bold text-center">
-          {{ month | monthName }}
-        </span>
-      </button>
-    </ng-container>
-  </div>
-</div>
